App_Deployment/requirements.txt
Pillow==6.2.0
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
dash==1.9.0
pandas==0.25.1
gunicorn==20.0.4
matplotlib==3.1.1
plotly-express==0.4.1
plotly==4.5.1
psycopg2==2.8.5
scikit-learn==0.21.3
urllib3==1.24.2
wordcloud==1.7.0
